+++ commons/ava_testsuite/rpc_workflow_test/rpc_workflow_test_.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_networks"
 	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_services"
+	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_testsuite/rpc_workflow_runner"
 	"github.com/kurtosis-tech/kurtosis/commons/networks"
 	"github.com/kurtosis-tech/kurtosis/commons/testsuite"
 	"github.com/palantir/stacktrace"
@@ -21,10 +22,11 @@ const (
 	stakeAmount       = int64(30000000000000)
 	delegatorAmount   = int64(30000000000000)
 
-	regularNodeServiceId   = 0
-	delegatorNodeServiceId = 1
+	regularNodeServiceId   networks.ServiceID = "validator-node"
+	delegatorNodeServiceId networks.ServiceID = "delegator-node"
 
-	normalNodeConfigId = 0
+	networkAcceptanceTimeoutRatio = 0.3
+	normalNodeConfigId networks.ConfigurationID = 0
 )
 
 type StakingNetworkRpcWorkflowTest struct {
@@ -32,8 +34,11 @@ type StakingNetworkRpcWorkflowTest struct {
 }
 
 func (test StakingNetworkRpcWorkflowTest) Run(network networks.Network, context testsuite.TestContext) {
+	// =============================== SETUP GECKO CLIENTS ======================================
 	castedNetwork := network.(ava_networks.TestGeckoNetwork)
 	stakerClient, err := castedNetwork.GetGeckoClient(regularNodeServiceId)
+	networkAcceptanceTimeout := time.Duration(networkAcceptanceTimeoutRatio * float64(test.GetExecutionTimeout().Nanoseconds()))
+
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not get staker client"))
 	}
@@ -49,14 +54,18 @@ func (test StakingNetworkRpcWorkflowTest) Run(network networks.Network, context 
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not get delegator node ID."))
 	}
-	highLevelStakerClient := ava_networks.NewHighLevelGeckoClient(
+	highLevelStakerClient := rpc_workflow_runner.NewRpcWorkflowRunner(
 		stakerClient,
 		stakerUsername,
-		stakerPassword)
-	highLevelDelegatorClient := ava_networks.NewHighLevelGeckoClient(
+		stakerPassword,
+		networkAcceptanceTimeout)
+	highLevelDelegatorClient := rpc_workflow_runner.NewRpcWorkflowRunner(
 		delegatorClient,
 		delegatorUsername,
-		delegatorPassword)
+		delegatorPassword,
+		networkAcceptanceTimeout)
+  
+	// ====================================== ADD VALIDATOR ===============================
 	stakerXchainAddress, err := highLevelStakerClient.CreateAndSeedXChainAccountFromGenesis(seedAmount)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not seed XChain account from Genesis."))
@@ -78,15 +87,18 @@ func (test StakingNetworkRpcWorkflowTest) Run(network networks.Network, context 
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not add staker %s to default subnet.", stakerNodeId))
 	}
+
+	// ====================================== VERIFY NETWORK STATE ===============================
 	currentStakers, err := stakerClient.PChainApi().GetCurrentValidators(nil)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not get current stakers."))
 	}
-	logrus.Debugf("Number of current stakers: %d", len(currentStakers))
+	logrus.Debugf("Number of current validators: %d", len(currentStakers))
 	actualNumStakers := len(currentStakers)
 	expectedNumStakers := 6
 	context.AssertTrue(actualNumStakers == expectedNumStakers, stacktrace.NewError("Actual number of stakers, %v, != expected number of stakers, %v", actualNumStakers, expectedNumStakers))
-	// Adding delegators
+
+	// ========================= ADD DELEGATOR AND TRANSFER FUNDS TO XCHAIN ======================
 	err = highLevelDelegatorClient.AddDelegatorOnSubnet(stakerNodeId, delegatorPchainAddress, delegatorAmount)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not add delegator %s to default subnet.", delegatorNodeId))
@@ -101,7 +113,9 @@ func (test StakingNetworkRpcWorkflowTest) Run(network networks.Network, context 
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Failed to transfer Ava from PChain to XChain."))
 	}
-	xchainAccountInfo, err := stakerClient.XChainApi().GetBalance(stakerXchainAddress, ava_networks.AVA_ASSET_ID)
+
+	// ================================ VERIFY NETWORK STATE =====================================
+	xchainAccountInfo, err := stakerClient.XChainApi().GetBalance(stakerXchainAddress, rpc_workflow_runner.AVA_ASSET_ID)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Failed to get account info for account %v.", stakerXchainAddress))
 	}
@@ -111,14 +125,16 @@ func (test StakingNetworkRpcWorkflowTest) Run(network networks.Network, context 
 }
 
 func (test StakingNetworkRpcWorkflowTest) GetNetworkLoader() (networks.NetworkLoader, error) {
-	serviceConfigs := map[int]ava_networks.TestGeckoNetworkServiceConfig{
+	// Define possible service configurations.
+	serviceConfigs := map[networks.ConfigurationID]ava_networks.TestGeckoNetworkServiceConfig{
 		normalNodeConfigId: *ava_networks.NewTestGeckoNetworkServiceConfig(true, ava_services.LOG_LEVEL_DEBUG, test.ImageName, 2, 2, nil),
 	}
-	desiredServices := map[int]int{
+	// Define which services use which configurations.
+	desiredServices := map[networks.ServiceID]networks.ConfigurationID{
 		regularNodeServiceId:   normalNodeConfigId,
 		delegatorNodeServiceId: normalNodeConfigId,
 	}
-
+	// Return a Gecko test net with this service:configuration mapping.
 	return ava_networks.NewTestGeckoNetworkLoader(
 		true,
 		test.ImageName,
@@ -129,6 +145,11 @@ func (test StakingNetworkRpcWorkflowTest) GetNetworkLoader() (networks.NetworkLo
 		desiredServices)
 }
 
-func (test StakingNetworkRpcWorkflowTest) GetTimeout() time.Duration {
-	return 90 * time.Second
+func (test StakingNetworkRpcWorkflowTest) GetExecutionTimeout() time.Duration {
+	return 5 * time.Minute
+}
+
+func (test StakingNetworkRpcWorkflowTest) GetSetupBuffer() time.Duration {
+	// TODO drop this down when the availability checker doesn't have a sleep (becuase we spin up a bunch of nodes before the test starts executing)
+	return 6 * time.Minute
 }