+++ commons/ava_testsuite/unrequested_chit_spammer_test/unrequested_chit_spammer_test_.go
@@ -5,23 +5,28 @@ import (
 
 	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_networks"
 	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_services"
+	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_testsuite/rpc_workflow_runner"
 	"github.com/kurtosis-tech/kurtosis/commons/networks"
 	"github.com/kurtosis-tech/kurtosis/commons/testsuite"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
+	"strconv"
 )
 
 const (
-	normalNodeConfigId   = 1
-	byzantineConfigId    = 2
-	byzantineUsername    = "byzantine_gecko"
-	byzantinePassword    = "byzant1n3!"
-	byzantineChitSpammer = "chit-spammer"
-	stakerUsername       = "staker_gecko"
-	stakerPassword       = "test34test!23"
-	normalNodeServiceId  = 4
-	seedAmount           = int64(50000000000000)
-	stakeAmount          = int64(30000000000000)
+	normalNodeConfigId networks.ConfigurationID = 1
+	byzantineConfigId networks.ConfigurationID = 2
+	byzantineUsername = "byzantine_gecko"
+	byzantinePassword = "byzant1n3!"
+	stakerUsername = "staker_gecko"
+	stakerPassword = "test34test!23"
+	normalNodeServiceId networks.ServiceID = "normal-node"
+	byzantineNodePrefix string = "byzantine-node-"
+	numberOfByzantineNodes = 4
+	seedAmount               = int64(50000000000000)
+	stakeAmount              = int64(30000000000000)
+
+	networkAcceptanceTimeoutRatio = 0.3
 )
 
 // ================ Byzantine Test - Spamming Unrequested Chit Messages ===================================
@@ -32,15 +37,19 @@ type StakingNetworkUnrequestedChitSpammerTest struct {
 
 func (test StakingNetworkUnrequestedChitSpammerTest) Run(network networks.Network, context testsuite.TestContext) {
 	castedNetwork := network.(ava_networks.TestGeckoNetwork)
-	for i := 0; i < normalNodeServiceId; i++ {
-		byzClient, err := castedNetwork.GetGeckoClient(i)
+	networkAcceptanceTimeout := time.Duration(networkAcceptanceTimeoutRatio * float64(test.GetExecutionTimeout().Nanoseconds()))
+
+	// ============= ADD SET OF BYZANTINE NODES AS VALIDATORS ON THE NETWORK ===================
+	for i := 0; i < numberOfByzantineNodes; i++ {
+		byzClient, err := castedNetwork.GetGeckoClient(networks.ServiceID(byzantineNodePrefix + strconv.Itoa(i)))
 		if err != nil {
 			context.Fatal(stacktrace.Propagate(err, "Failed to get byzantine client."))
 		}
-		highLevelByzClient := ava_networks.NewHighLevelGeckoClient(
+		highLevelByzClient := rpc_workflow_runner.NewRpcWorkflowRunner(
 			byzClient,
 			byzantineUsername,
-			byzantinePassword)
+			byzantinePassword,
+			networkAcceptanceTimeout)
 		err = highLevelByzClient.GetFundsAndStartValidating(seedAmount, stakeAmount)
 		if err != nil {
 			context.Fatal(stacktrace.Propagate(err, "Failed add client as a validator."))
@@ -51,6 +60,8 @@ func (test StakingNetworkUnrequestedChitSpammerTest) Run(network networks.Networ
 		}
 		logrus.Debugf("Number of current stakers: %d", len(currentStakers))
 	}
+
+	// =================== ADD NORMAL NODE AS A VALIDATOR ON THE NETWORK =======================
 	availabilityChecker, err := castedNetwork.AddService(normalNodeConfigId, normalNodeServiceId)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Failed to add normal node with high quorum and sample to network."))
@@ -62,14 +73,17 @@ func (test StakingNetworkUnrequestedChitSpammerTest) Run(network networks.Networ
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Failed to get staker client."))
 	}
-	highLevelNormalClient := ava_networks.NewHighLevelGeckoClient(
+	highLevelNormalClient := rpc_workflow_runner.NewRpcWorkflowRunner(
 		normalClient,
 		stakerUsername,
-		stakerPassword)
+		stakerPassword,
+		networkAcceptanceTimeout)
 	err = highLevelNormalClient.GetFundsAndStartValidating(seedAmount, stakeAmount)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Failed add client as a validator."))
 	}
+
+	// ============= VALIDATE NETWORK STATE DESPITE BYZANTINE BEHAVIOR =========================
 	currentStakers, err := normalClient.PChainApi().GetCurrentValidators(nil)
 	if err != nil {
 		context.Fatal(stacktrace.Propagate(err, "Could not get current stakers."))
@@ -78,31 +92,16 @@ func (test StakingNetworkUnrequestedChitSpammerTest) Run(network networks.Networ
 	actualNumStakers := len(currentStakers)
 	expectedNumStakers := 10
 	context.AssertTrue(actualNumStakers == expectedNumStakers, stacktrace.NewError("Actual number of stakers, %v, != expected number of stakers, %v", actualNumStakers, expectedNumStakers))
+
 }
 
 func (test StakingNetworkUnrequestedChitSpammerTest) GetNetworkLoader() (networks.NetworkLoader, error) {
-	serviceIdConfigMap := map[int]int{}
-	for i := 0; i < normalNodeServiceId; i++ {
-		serviceIdConfigMap[i] = byzantineConfigId
-	}
-	return getByzantineNetworkLoader(serviceIdConfigMap, test.ByzantineImageName, test.NormalImageName)
-}
-func (test StakingNetworkUnrequestedChitSpammerTest) GetTimeout() time.Duration {
-	return 720 * time.Second
-}
-
-// =============== Helper functions =============================
-
-/*
-Args:
-	desiredServices: Mapping of service_id -> configuration_id for all services *in addition to the boot nodes* that the user wants
-*/
-func getByzantineNetworkLoader(desiredServices map[int]int, byzantineImageName string, normalImageName string) (networks.NetworkLoader, error) {
-	serviceConfigs := map[int]ava_networks.TestGeckoNetworkServiceConfig{
+	// Define normal node and byzantine node configurations
+	serviceConfigs := map[networks.ConfigurationID]ava_networks.TestGeckoNetworkServiceConfig{
 		normalNodeConfigId: *ava_networks.NewTestGeckoNetworkServiceConfig(
 			true,
 			ava_services.LOG_LEVEL_DEBUG,
-			normalImageName,
+			test.NormalImageName,
 			6,
 			8,
 			make(map[string]string),
@@ -110,21 +109,37 @@ func getByzantineNetworkLoader(desiredServices map[int]int, byzantineImageName s
 		byzantineConfigId: *ava_networks.NewTestGeckoNetworkServiceConfig(
 			true,
 			ava_services.LOG_LEVEL_DEBUG,
-			byzantineImageName,
+			test.ByzantineImageName,
 			2,
 			2,
-			map[string]string{"byzantine-behavior": byzantineChitSpammer},
+			map[string]string{"byzantine-behavior": "chit-spammer"},
 		),
 	}
-	logrus.Debugf("Byzantine Image Name: %s", byzantineImageName)
-	logrus.Debugf("Normal Image Name: %s", normalImageName)
+
+	// Define the map from service->configuration for the network
+	serviceIdConfigMap := map[networks.ServiceID]networks.ConfigurationID{}
+	for i := 0; i < numberOfByzantineNodes; i++ {
+		serviceIdConfigMap[networks.ServiceID(byzantineNodePrefix + strconv.Itoa(i))] = byzantineConfigId
+	}
+	logrus.Debugf("Byzantine Image Name: %s", test.ByzantineImageName)
+	logrus.Debugf("Normal Image Name: %s", test.NormalImageName)
 
 	return ava_networks.NewTestGeckoNetworkLoader(
 		true,
-		normalImageName,
+		test.NormalImageName,
 		ava_services.LOG_LEVEL_DEBUG,
 		2,
 		2,
 		serviceConfigs,
-		desiredServices)
+		serviceIdConfigMap)
+}
+
+func (test StakingNetworkUnrequestedChitSpammerTest) GetExecutionTimeout() time.Duration {
+	return 5 * time.Minute
+}
+
+func (test StakingNetworkUnrequestedChitSpammerTest) GetSetupBuffer() time.Duration {
+	// TODO drop this when the availabilityChecker doesn't have a sleep, because we spin up a *bunch* of nodes before test
+	//  execution starts
+	return 12 * time.Minute
 }