+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # TBD
 * Split `staking_network_tests` into separate files per test
 * Upgrade to Kurtosis version with simplified service config definition
+* Implement network consensus timeouts in high level gecko client as fractions of total test timeout
+* Add CI checks to make sure changelog is updated
+* Upgrade to Kurtosis version using custom structs for service/config IDs (rather than ints)
+* Significantly up test execution timeouts
+* Make controller Docker image `tee` to the logfile, rather than redirecting all output
+* Implement a 10-second HTTP request timeout in the Gecko client
+* Upgraded to Kurtosis that allows the setup buffer to be configured at a per-test level, and gave a generous setup buffer for as long as the Gecko availability checker core has a `time.Sleep` in it
+* Rename HighLevelGeckoClient to RpcWorkflowRunner and move to ava_testsuite package
+* Bugfix on Travis CI checks for CHANGELOG entries
+* Add comments and documentation in RPC workflow and chit spammer tests
+* Using network ID in controller image rather than network name
+* Changing service IDs to strings rather than integers
+* Update function interface for GetStartCommand according to net.IP change in kurtosis
+* Added a script to clean Docker environment
 
 # 0.6.0
 * Use Kurtosis version that allows the user to configure network width